requirements.txt
streamlit==1.28.1
requests==2.31.0
beautifulsoup4==4.12.2
sentence-transformers==2.2.2
chromadb==0.4.18
langchain==0.1.0
langchain-openai==0.0.5
openai==1.6.1
python-dotenv==1.0.0
tiktoken==0.5.2
